requirements.txt
# Core dependencies
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.0
pyyaml>=5.4.0

# Data processing
opencv-python>=4.5.0
h5py>=3.1.0
imageio>=2.9.0
scikit-video>=1.1.11

# Utils
tqdm>=4.60.0
matplotlib>=3.3.0
